refactor: Extract influxdb_iox remote CLI tests to their own file
