+++ influxdb_iox/tests/end_to_end_cases/mod.rs
@@ -11,5 +11,31 @@ mod logging;
 mod metrics;
 mod namespace;
 mod querier;
+mod remote;
 mod schema;
 mod tracing;
+
+/// extracts the parquet filename from JSON that looks like
+/// ```text
+/// {
+///    "id": "1",
+///    ...
+//     "objectStoreId": "fa6cdcd1-cbc2-4fb7-8b51-4773079124dd",
+///    ...
+/// }
+/// ```
+fn get_object_store_id(output: &[u8]) -> String {
+    let v: serde_json::Value = serde_json::from_slice(output).unwrap();
+    // We only process the first value, so panic if it isn't there
+    let arr = v.as_array().unwrap();
+    assert_eq!(arr.len(), 1);
+    let id = arr[0]
+        .as_object()
+        .unwrap()
+        .get("objectStoreId")
+        .unwrap()
+        .as_str()
+        .unwrap();
+
+    id.to_string()
+}
